Refactor configuration step 1.
